feat: address entity and patient entity features added

+++ src/main/java/com/yasmin/projects/medsystem/api/service/DoctorService.java
@@ -3,7 +3,6 @@ package com.yasmin.projects.medsystem.api.service;
 import com.yasmin.projects.medsystem.api.domain.doctor.*;
 import com.yasmin.projects.medsystem.api.repository.DoctorRepository;
 import jakarta.persistence.EntityNotFoundException;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -11,7 +10,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 
 @Service
-public class  DoctorService {
+public class DoctorService {
 
     private final DoctorRepository doctorRepository;
     private final SpecialityService specialityService;
@@ -24,38 +23,39 @@ public class  DoctorService {
     }
 
     public ResponseDoctorInfoDTO saveDoctor(RequestDoctorInfoDTO doctorInfoDTO) {
-        Doctor doctor;
-        if (doctorInfoDTO != null){
-            List<Speciality> specialities = specialityService.getSpecialities(doctorInfoDTO.specialities());
-            List<WorkingHour> workingHours =  workingHourService.saveWorkingHours(doctorInfoDTO.workingHours());
+        List<Speciality> specialities = specialityService.getSpecialities(doctorInfoDTO.specialities());
+        List<WorkingHour> workingHours = workingHourService.saveWorkingHours(doctorInfoDTO.workingHours());
 
-            doctor = doctorRepository.save(new Doctor(doctorInfoDTO, specialities, workingHours));
+        Doctor doctor = doctorRepository.save(new Doctor(doctorInfoDTO, specialities, workingHours));
 
-            for(WorkingHour workingHourElement : workingHours){
-                workingHourElement.setDoctor(doctor);
-            }
-
-            return new ResponseDoctorInfoDTO(doctor);
+        for (WorkingHour workingHourElement : workingHours) {
+            workingHourElement.setDoctor(doctor);
         }
 
-        throw  new DataIntegrityViolationException("Crm ou email já cadastrado!!");
+        return new ResponseDoctorInfoDTO(doctor);
     }
 
-    public Page<ResponseDoctorInfoDTO> getAllDoctors(Pageable pageable){
+    public Page<ResponseDoctorInfoDTO> getAllDoctors(Pageable pageable) {
         return doctorRepository.findAll(pageable).map(ResponseDoctorInfoDTO::new);
     }
 
-    public ResponseDoctorInfoDTO getDoctorById(Integer id){
+    public ResponseDoctorInfoDTO getDoctorById(Integer id) {
         return doctorRepository.findById(id).map(ResponseDoctorInfoDTO::new)
-                .orElseThrow(() -> new EntityNotFoundException("Doutor com não encontrado!!"));
+                .orElseThrow(() -> new EntityNotFoundException("Doutor não encontrado!!"));
     }
 
-    public ResponseDoctorInfoDTO updateDoctor(RequestUpdateDoctorDTO doctorInfoDTO, Integer id){
+    public ResponseDoctorInfoDTO updateDoctor(RequestUpdateDoctorDTO doctorInfoDTO, Integer id) {
+        if(!doctorRepository.existsById(id)){
+            throw new EntityNotFoundException("Doutor não encontrado!!");
+        }
         Doctor doctor = doctorRepository.getReferenceById(id);
         return new ResponseDoctorInfoDTO(doctorRepository.save(doctor.update(doctorInfoDTO)));
     }
 
-    public void deleteDoctor(Integer id){
+    public void deleteDoctor(Integer id) {
+        if(!doctorRepository.existsById(id)){
+            throw new EntityNotFoundException("Doutor não encontrado!!");
+        }
         doctorRepository.deleteById(id);
     }
 }
